requirements.txt
astroid==2.4.2
Django==2.2.16
django-filter==2.3.0
django-phonenumber-field==5.0.0
djangorestframework==3.11.1
djangorestframework-jsonapi==3.2.0
inflection==0.5.1
isort==5.5.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumbers==8.12.9
Pillow==7.2.0
psycopg2-binary==2.8.5
pylint==2.6.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
